Move struct definition to ty module.
