+++ sway-core/src/control_flow_analysis/dead_code_analysis.rs
@@ -5,8 +5,8 @@ use crate::{
     semantic_analysis::{
         ast_node::{
             TyAbiDeclaration, TyCodeBlock, TyConstantDeclaration, TyEnumDeclaration,
-            TyFunctionDeclaration, TyStructDeclaration, TyStructExpressionField,
-            TyTraitDeclaration, TyVariableDeclaration, VariableMutability,
+            TyFunctionDeclaration, TyStructExpressionField, TyTraitDeclaration,
+            TyVariableDeclaration, VariableMutability,
         },
         TyAsmRegisterDeclaration, TyAstNode, TyAstNodeContent, TyImplTrait,
         TyIntrinsicFunctionKind, TyStorageDeclaration,
@@ -375,12 +375,12 @@ fn connect_declaration(
 /// Connect each individual struct field, and when that field is accessed in a subfield expression,
 /// connect that field.
 fn connect_struct_declaration(
-    struct_decl: &TyStructDeclaration,
+    struct_decl: &ty::TyStructDeclaration,
     graph: &mut ControlFlowGraph,
     entry_node: NodeIndex,
     tree_type: &TreeType,
 ) {
-    let TyStructDeclaration {
+    let ty::TyStructDeclaration {
         name,
         fields,
         visibility,
