+++ sway-core/src/declaration_engine/declaration_wrapper.rs
@@ -4,9 +4,10 @@ use sway_error::error::CompileError;
 use sway_types::Span;
 
 use crate::{
+    language::ty,
     semantic_analysis::{
         TyAbiDeclaration, TyConstantDeclaration, TyEnumDeclaration, TyImplTrait,
-        TyStorageDeclaration, TyStructDeclaration, TyTraitDeclaration, TyTraitFn,
+        TyStorageDeclaration, TyTraitDeclaration, TyTraitFn,
     },
     type_system::{CopyTypes, TypeMapping},
     TyFunctionDeclaration,
@@ -22,7 +23,7 @@ pub(crate) enum DeclarationWrapper {
     Trait(TyTraitDeclaration),
     TraitFn(TyTraitFn),
     ImplTrait(TyImplTrait),
-    Struct(TyStructDeclaration),
+    Struct(ty::TyStructDeclaration),
     Storage(TyStorageDeclaration),
     Abi(TyAbiDeclaration),
     Constant(Box<TyConstantDeclaration>),
@@ -155,7 +156,10 @@ impl DeclarationWrapper {
         }
     }
 
-    pub(super) fn expect_struct(self, span: &Span) -> Result<TyStructDeclaration, CompileError> {
+    pub(super) fn expect_struct(
+        self,
+        span: &Span,
+    ) -> Result<ty::TyStructDeclaration, CompileError> {
         match self {
             DeclarationWrapper::Struct(decl) => Ok(decl),
             DeclarationWrapper::Unknown => Err(CompileError::Internal(
