+++ sway-core/src/language/ty/declaration/struct.rs
@@ -1,0 +1,64 @@
+use sway_types::{Ident, Span, Spanned};
+
+use crate::{
+    language::Visibility, semantic_analysis::TyStructField, type_system::*, AttributesMap,
+};
+
+#[derive(Clone, Debug, Eq)]
+pub struct TyStructDeclaration {
+    pub name: Ident,
+    pub fields: Vec<TyStructField>,
+    pub type_parameters: Vec<TypeParameter>,
+    pub visibility: Visibility,
+    pub(crate) span: Span,
+    pub attributes: AttributesMap,
+}
+
+// NOTE: Hash and PartialEq must uphold the invariant:
+// k1 == k2 -> hash(k1) == hash(k2)
+// https://doc.rust-lang.org/std/collections/struct.HashMap.html
+impl PartialEq for TyStructDeclaration {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+            && self.fields == other.fields
+            && self.type_parameters == other.type_parameters
+            && self.visibility == other.visibility
+    }
+}
+
+impl CopyTypes for TyStructDeclaration {
+    fn copy_types(&mut self, type_mapping: &TypeMapping) {
+        self.fields
+            .iter_mut()
+            .for_each(|x| x.copy_types(type_mapping));
+        self.type_parameters
+            .iter_mut()
+            .for_each(|x| x.copy_types(type_mapping));
+    }
+}
+
+impl CreateTypeId for TyStructDeclaration {
+    fn create_type_id(&self) -> TypeId {
+        insert_type(TypeInfo::Struct {
+            name: self.name.clone(),
+            fields: self.fields.clone(),
+            type_parameters: self.type_parameters.clone(),
+        })
+    }
+}
+
+impl Spanned for TyStructDeclaration {
+    fn span(&self) -> Span {
+        self.span.clone()
+    }
+}
+
+impl MonomorphizeHelper for TyStructDeclaration {
+    fn type_parameters(&self) -> &[TypeParameter] {
+        &self.type_parameters
+    }
+
+    fn name(&self) -> &Ident {
+        &self.name
+    }
+}
